The language requiring DESTROY_CLIENTID to destroy any existing state contradicts language in 5661, which requires failure in the presence of such state:
	"If there are sessions (both idle and non-idle), opens, locks,
	delegations, layouts, and/or wants (Section 18.49) associated
	with the unexpired lease of the client ID, the server MUST
	return NFS4ERR_CLIENTID_BUSY."

That was a mistake; we'd rather DESTROY_CLIENTID continues to function
as it it did in 4.1.  The above language was probably meant to apply to
the client-reboot case of EXCHANGE_ID/CREATE_SESSION, but that's already
covered by a previous paragraph.  So just remove this language.

Also clarify that it's the server reply that governs whether the new
behavior here is in effect.

Signed-off-by: loghyr <loghyr@gmail.com>