NFSv4.2 issues list
On Thu, Apr 24, 2014 at 12:00:18PM -0700, Tom Haynes wrote:
> 1) space_reserved flag - what does it mean now that reservations
>    are byte ranges?
>
>    See the recent chatter on the WG.

Here is my proposal to address this issue:

The terminology of allocated vs unallocated blocks in READ_PLUS and SEEK
is confusing as there it doesn't have a whole lot of meaning for a remote
filesystem, especially if it sits on top of complex COW or snapshotting
filesystem.  Even if a block is allocated writes into it might return NOSPC,
and even if it's not allocated it might have space reserved making writes
into it semaingless.

So intead of returning the allocated/unallocated information return the
reservation status of a hole or data block range, as that actually is
meaninful to the client.  For a traditional, simple Unix filesystem the
two are equivalent, but the the strictly defined meaning of reserved
vs unreserved actually does matter for more complex server implementations.

Because we want the answert to be exact but servers might not always be able
to get the exact information add a third possible unknown state to the
reservation status.  This should also help with the earlier discussions if
the allocation status should be exact or a hint.

Signed-off-by: loghyr <loghyr@primarydata.com>