NFSv4.2 issues list
On Thu, Apr 24, 2014 at 12:00:18PM -0700, Tom Haynes wrote:
> 2) Do we want to maintain punch hole versus write zeros OR do we
>    want to leave that to the server?
>
>    We added this, but is it realistic?
>
>    Also, is it a hint or mandatory?
>
>    Either way, I want us to revisit it.

Hole punching vs zeroing is an important difference in semantics, and it
becomes even more important with the strictly defined space reservation
semantics.

I've attached a patch (ontop of my previous patch in this thread) which
tries to address these issues by splitting them up into different operations.

Zeroing a file of parts of it can be done with the WRITE_SAME operation
in the current ADB form - an ADB without pattern or block number will
simply zero the file, so that part of WRITE_HOLE is properly replaces by
WRITE_SAME already, I've just added an example section for it.

The hole punch now becomes a simple metadata operation mirroring ALLOCATE,
thus termed 'DEALLOCATE', and it now actually drops the reservation, which
is the only sensible semantic for it to implement the sort of hole punches
clients expect.  It also is purely synchronous as it just is a generalization
of SETATTR to a smaller size, which servers must handle synchronously already.

Signed-off-by: loghyr <loghyr@primarydata.com>