Have the OFFLOAD_STATUS be generic
Built on top of the OFFLOAD_CANCEL patch I just sent out.

This straightens out the text in OFFLOAD_STATUS to focus
on the generic need and not the COPY speficics.

Also, all of the verbage about REQUIREMENTs for which
op is needed is already given in the main COPY overview,
so axe that stuff out!

Signed-off-by: Tom Haynes <Thomas.Haynes@primarydata.com>