COPY: intra-server copies also require a valid src stateid
Without a valid src stateid the server must perform an implicit
open, which would not be able to respect locking or delegation.

Additionally at least the Linux server requires the stateid to
easily lookup the file object, as demonstrated by the COPY
prototype implementation which would fail to copy with an invalid
src stateid.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Tom Haynes <loghyr@primarydata.com>